Make game `owner` into a required field
Update game model making owner a required field to ensure that every game always has a user to manage it. Generate the corresponding migration script.

Add a pre-migration operation to the migration script that deletes existing games that have no owner, this ensures that the change will not be breaking to an existing instance.