Isolate the token and username getters and setters
Isolate the token and username getters and setters so that it will be easier to implement React contexts in a future commit.

Create and implement the `useLocalStorage` function for these getters and setters to reduce code repetition. This function is an improvement over the previous usage of `setAuthData` because now it is impossible to set tokens and username without also updating local storage.