Add a context provider for auth members (#67)
* Isolate the token and username getters and setters

Isolate the token and username getters and setters so that it will be easier to implement React contexts in a future commit.

Create and implement the `useLocalStorage` function for these getters and setters to reduce code repetition. This function is an improvement over the previous usage of `setAuthData` because now it is impossible to set tokens and username without also updating local storage.

* Create auth context for auth getters and setters

Create a React context called `AuthContext` to provide getters and setters for tokens and username to components that need them, without passing these getters and setters down the component tree.

Implement `AuthContext` everywhere that interacts with tokens and username. The App component now looks much more readable without the getters and setters being passed to it's child components.