Combine both user viewsets into a single viewset
Combine both user view sets into a single view set that uses methods to determine the query set and serializer based on the URI. The URI pattern now contains the plural "users" instead of "user". The URI can be optionally appended with `<username>/` to show details about one user, as long as that user is the request issuer. This resource naming pattern is more conventional for REST APIs.