Make the API more conventional (#70)
* Combine both user viewsets into a single viewset

Combine both user view sets into a single view set that uses methods to determine the query set and serializer based on the URI. The URI pattern now contains the plural "users" instead of "user". The URI can be optionally appended with `<username>/` to show details about one user, as long as that user is the request issuer. This resource naming pattern is more conventional for REST APIs.

* Rename and reorder API paths, names and imports

Rename and reorder API paths, names and imports to improve readability of the code.

* Set token serializer field type to password

Set the obtain token by email serializer password field type to password, so it's corresponding form looks better in the web browsable API.

* Add newline to end of URL file

Add a newline to the end of the URL file because that is the conventional style.

* Make URI patterns and route names conventional

Pluralize the token URI pattern and singularize the names of the games and users routes. This change makes the names more conventional.