Add graceful handling of session expiration
When a session expires, or the user isn't signed in, but they are trying to use a signed-in-only page, an alternative version of the sign in dialog opens. The user can sign in to continue using the page, or return to home as a guest.

Also add a blur effect with fade in/out to the backdrop.