Graceful handling of refresh token expiration (automatic sign out) (#77)
* Add CSR backup fetch if SSR fetch fails

For the game list page, make the component do a client-side fetch if the server-side fetch could have failed. This ensures that if the auth tokens have expired, the client still clears it's tokens and signs the user out.

* Rename some variables in the games sort

* Fix text on the home page

* Rename the "helpers" folder to "functions"

* Set custom `<title>`s for each page

* Give the context provider a more general name

Rename the context provider, currently used only for username and tokens, from "AuthContext" to "ContextProvider". This sets up the provider to take on the role of a more general provider in the future.

* Revert "Give the context provider a more general name"

This reverts commit 4d82dbf55b10d01aef0c5c3ce87fa4e98d638fa1.

* Wrap the auth provider inside the root provider

* Move dialog state into new dialog provider

Move the state responsible for tracking which dialog is open, `dialog`, into a new provider called `DialogProvider`, which provides `useDialogContext`. Within the scope of this commit, it's just a refactor. But going forward, the dialog state will be accessible anywhere, which is *nice*.

* Add graceful handling of session expiration

When a session expires, or the user isn't signed in, but they are trying to use a signed-in-only page, an alternative version of the sign in dialog opens. The user can sign in to continue using the page, or return to home as a guest.

Also add a blur effect with fade in/out to the backdrop.

* Add focus trap to trap focus for modal dialogs

* Improve ref types used in the button component

* Enable modal dialog closure using the ESC key

* Make the sign in button focusable

Make the sign in button within the sign in modal dialog focusable, because it wasn't before, unlike the other fields and cancel button.

* Add `aria-modal` to improve accessibility

* Rename dialogs to modals

Rename `dialog` to `modal` because they are specifically modals, which is a subset of dialogs.

* Add modal X (close) button

* Make everything behind modal inert

* Fix senator inspector issue

Some reason the .senator class wasn't explicitly absolute

* Be explicit on token lifetimes