Add game participants (#95)
* Add a game participant model

* Add a migration for the new game participant model

* Add game participants to the admin site

* Make owner a game participant on game creation

Make the API view set always add a new game participant on game creation, with the owner as the participant. This is a current design choice that could be changed in future without needing to alter the model, because the view set is responsible for this behavior.

* Add participants to game read API

Change the game read serializer to show participants in read API responses

* Make Django admin game page show participants

Change the Django admin editor for the game model so that it lists participants.

* Restore commented out code

This code is used to ensure that the not authorized page is shown if the user is not signed in.

* Make participants show up in the game page

Make participants show up in the game page in a comma separated list. Also update the frontend Game class to expect a list of strings as the participants.

* Add participant count to list page

Add a participant count as a new column in the game list.

* Clean up whitespace

Clean up whitespace and also add a game participant count check on the game detail page.