Replace the fix with a better fix
Move the font variable declarations to the `<head>`'s `<style>` so that they are exposed to all elements, including MUI DataGrid popovers and menus. The fonts used in the project are now only accessible via the their CSS variables.

This solution addresses the root problem better than the previous commit did.