Try using a Docker compose file
Try adding a Docker compose, containing the database and the Django app, and using it during the backend CI workflow instead of a Docker Django app and a non-Docker database. This may allow the Django app to connect to the database, because it has been failing to connect to the database in recent CI runs.