Migrate to Docker (#121)
* Create a Dockerfile

* Update the CI workflow for the backend

Update the CI workflow for the backend to use a Docker environment instead of a Python environment

* Try to fix Docker container DB connection issue

Try changing the `RDS_HOSTNAME` so the Docker container can access the database, because it failed to connect to the database after the last commit.

* Try using a Docker compose file

Try adding a Docker compose, containing the database and the Django app, and using it during the backend CI workflow instead of a Docker Django app and a non-Docker database. This may allow the Django app to connect to the database, because it has been failing to connect to the database in recent CI runs.

* Move some Django configs to the Dockerfile

Move some Django deployment configurations from .ebextensions to the Dockerfile because these configurations may not be applicable to a Docker production environment.

* Move deployment configs to shell script

Try moving deployment configurations to a shell script.

* Rename the Docker Compose file

Rename the Docker Compose file to "docker-compose.test.yml", which corresponds to it's sole usage within the test environment.

* Add a Docker Compose for local development

* Try removing the `entrypoint` shell script

* Add a docker run file and remove configs

Add a docker run file for deployment and remove the elastic beanstalk and NGINX configs (possibly a temporary measure) to simplify deployment.

* Add a start app command to the Dockerfile

Add a command that starts the app to the Dockerfile. This is essential for the app to start in production, so it's absence was making successful deployment impossible.

* Add temp hostname to allowed hosts

Add the temporary hostname of the production environment to the allowed hosts list in the Django settings file.

* Revert "Add temp hostname to allowed hosts"

This reverts commit 84be113b13ea5322aa8b8f6ff0382ab5df6e56ce.

* Add `localhost` to allowed hosts

* Allow access to a private DockerHub repo

* Attempt to setup SSL/TLS

* Fix some issues with Docker deployment files

* Revert "Reorder 2 steps in the backend deployment workflow"

This reverts commit aa2bc5846cd9757a09cfbae5f204f1fc9e1afbef.

* Revert "Revert "Reorder 2 steps in the backend deployment workflow""

This reverts commit 246947113096e431e6b10e6d1fda601be7668b92.

* Move the NGINX config file

* Change ports in Dockerrun.aws.json

* Fix a mistake in the NGINX config

* Remove unnecessary file copy from Dockerfile

* Update Dockerfile

* Update SSL credential retrieval to use the AWS CLI

* Update Dockerfile

* Reinstate the installation of AWS CLI

* Fix server_name in NGINX configuration

* Enable Django logging

* Add NGINX and Gunicorn to local dev compose

* Attempt to fix the static files issue

* Add static volume and Django debug toolbar

* Fix debug environment variable type

* Attempt to fix port issue in CI

* Attempt to fix env var issue in CI

* Remove the "run migrations" step from backend CI

* Fix the static file issue

* Fix NGINX static files production configuration

* Fix admin site

* Fix local port issue

* Add database console logging

* Add a 120 seconds timeout to Gunicorn

* Setup Gunicorn logging

* Set permissions for Gunicorn logging directory

* Remove the Gunicorn logging files

* Remove DB logging and set console logging to DEBUG

Remove DB logging because it never worked in the first place, and change the console logging from "WARNING" to "DEBUG".

* Add another Django logger

* Remove manual dispatch for backend deployment

* Set Django log level to WARNING

Change Django log level from DEBUG to WARNING, which is a more reasonable level.

* Reenable local use of the Django dev server