Add start and play buttons with basic game page
Add start and play buttons to the game lobby page. Start is for the host to start the game, a one time action. Play is for anyone to join the game once it has been started. Add a basic placeholder game page, which is where you end up after pressing either of these buttons.