Fix missing email address on account page
Fix a bug where the email address wasn't showing up on the account page. This fix expanded into a change in how user data is stored: instead of storing the username, the User object is now serialized and stored in a cookie. The way in which data is deserialized was also generalized, which is nice.