Use normalization for state in the game play page (#155)
* Used normalized state storage in game play page

Introduce a generic `Collection` class for storing state in a normalized structure. This is used in the play game page for family senators, factions and game participants. The host user is now prefetched with the game instance, however this is not the default behavior of the API and must be specified via the "prefetch_user" URL parameter.

* Fix frontend build errors