Populate meta section and play page synch
Populate the meta section with some progress info and a button to return to the lobby. Replicate the lobby page's synchronization functionality in the play page. Add a spinner to the play page to appear whilst game data is loading. Fix a consul icon bug where the linear relationship between portrait and icon size was broken. Update Font Awesome package versions.