Fix a horrible performance issue
Fix a horrible performance issue where the page would have frame drops and scroll lag when using the senator list. I think this was caused by the browser having to paint too much content, but I'm not sure. The solution was to use a package called `react-virtualized` that only renders visible items in the list.