Make sorting and grouping persistent (#215)
* Make sorting and grouping persistent

Make sorting and grouping persistent by storing the relevant state on the game page. This is optional though - other instances of the `SenatorList` component default to storing this state themselves (in which case the sorting is not persistent).

* Make hooks always defined

Fix issue that was causing build failure: the new hooks were being conditionally defined, which worked, but apparently breaks a fundamental rule in React. Instead, the hooks must be defined, but their conditional usage is allowed.