Force re-render of portrait on senator prop change (#247)
Use the React `key` prop to force a re-render of the senator portrait component whenever the `senator` prop is changed. Also make the senator's Next.js `Image` component use a blurry placeholder whilst the full image loads. These two changes help improve the look of the senator images when using the app for the first time or with cache disabled.