Fix indefinite loading bug on game pages
Fix an issue where the game lobby and game play pages would continue loading indefinitely if the user's auth JWT tokens have expired. This was happening because the client had no way of knowing that the user's tokens had expired without making a HTTP request. Now there is a `fetchGame` call within in the `useWebSocket`'s `onClose` that will sign the user out if the call fails due to bad credentials.