Add `rank` attributes to senator & faction models
Add `rank` attributes to senator and faction models. Senator rank is 0 for HRAO then counting up in order of succession. Faction rank is 0 for the faction containing the HRAO then counting up in order of position (red -> yellow -> blue -> orange -> green -> violet -> red again etc).

Add a function that assigns rank to instances of these models. Call a migration-suitable version of this function in the migration script. Call the normal function in the mortality resolution, because that's when ranks may change.

Make the faction list a bit more similar to the senator list. Factions are now ordered by rank in the faction list and the senator list (when grouped by faction).