Enable frontend to build an action list
Make the changes that enable the frontend to build a list of action logs, separate from the notification list, for showing action logs for each senator in the history tab. At this stage the listed actions don't look very informative. Action logs for a given senator are fetched whenever the user opens the senator detail section for that senator. They are not fetched again until the game is fully reloaded, but new log items are passed by WebSocket messages just as new notifications come in.