Senator history tab (#267)
* Prepare notification data

Prepare notification data for use as a basis for creating senator-notification relations.

* Add senator-notification relation

Add a senator-notification relation model for tracking relationships between senators and notifications.

Fix a bug where the mortality resolution rank assignment was failing. Also fix a bug in mortality notification appearance.

* Rename Notification to ActionLog in the backend

Rename `Notification` to `ActionLog` in the backend, because it will be used for logging as well as for notifications. The thing called a "notification" in the frontend has not been renamed because it's still being used only for notifications, for now.

* Allow action logs to be filtered by senator

Update the action log view so action logs can be filtered by senator, using the senator-action log relation. Fix migration script 0034 by stopping it from deleting all `SenatorActionLog` data. Update existing model `__str__` functions to include instance ID, making the admin site easier to use.

* Set maximum widths for sections

Set maximum widths for the detail and progress sections.

* Add senator details tab selectors

* Rename the `Notification` class in the frontend

Rename the `Notification` class in the frontend to `ActionLog`, but leave variables and state that use it as-is, because those still use the `ActionLog` items as notifications.

* Enable frontend to build an action list

Make the changes that enable the frontend to build a list of action logs, separate from the notification list, for showing action logs for each senator in the history tab. At this stage the listed actions don't look very informative. Action logs for a given senator are fetched whenever the user opens the senator detail section for that senator. They are not fetched again until the game is fully reloaded, but new log items are passed by WebSocket messages just as new notifications come in.

* Fix game setup ranking bug

Fix a game setup bug where the senators and factions were not getting ranked during the initial game setup.

* Fix React and HTML rule violations

* Add Temporary Rome Consul notification

Add a notification at the start of the game to declare the Temporary Rome Consul. Also simplify the existing notification components.

* Make the history tab look better

Make the history tab look better. The tab now contains notification components, exactly like those in the notification section. However, these components have now been renamed to include the name "ActionLog" to reflect their dual/general purpose.

* Fix faction tab display issues

* Fix action log duplication issue

Fix an issue where duplicate action logs were appearing in the history tab.

* Past tense action log messages

Added past tense versions of all action log messages, for the senator history tab.

* Add no history placeholder

Add a placeholder for senators with no related action logs (i.e. no "history").

* Fix display issue with the lobby button

* Use Rome Consul icon for action log

Set the icon for the Temporary Rome Consul action log to the Rome Consul icon.

* Fix frontend build warnings

* Fix image alt values for accessibility

* Prevent creation of duplicate Temp RC action logs