Make dead senators have more info
Make dead senators show more info, specifically by making the senator's portrait, list item and details show former faction and titles. After death, senators are now considered to remain "in" the faction that they were in when they died - which is a bit weird but easier than making a new model to represent the former allegiance of dead senators.