Style changes (#299)
* Change styling of various parts of the game UI

* Fix game page layout issue

Fix a layout issue on the game page where the detail section would grow larger than one quarter of the full width if it's content was large.

* Introduce a new color scheme

Introduce a new color scheme. The colors can be viewed at `/colors`. The colors have been chosen but not used much in the site yet, except stone and white for backgrounds.

* Rework landing page, top bar and sign out dialog

* Fix up styles in various pages

* Fix game title sizes and colors in the lobby page