Fix mortality/forum phase indexing issue
Fix phase indexing issue that was causing a bug where the UI showed the current phase as mortality when it should show faction. This was caused by the index of all phases, except the initial faction phase, being set to 1, making it impossible for parts of the system to figure out which phase is the "latest".