Change attribute styles to match the senator count
Change the style of attribute grid values to emphasize values by making the font larger and semi bold. Tweak the layout to fit the larger values. Remove the "Total" text from faction attribute names because it looks a bit cleaner. These changes were partially made so that the attribute grid style matches that of the senator list count.