Add a new backend cleanup cron job (builder prune)
Update the backend cron configuration to include a new cron job that runs a docker builder prune command every 3 hours, except 6am because that's when the existing system prune cron job runs. The builder prune removes unused build caches, which is a subset of what system prune removes. This new job will prevent accumulation of cached builds if several commits are being pushed to `main` in a single day.