Introduce the public functions module
Turn the functions folder in the backend into a Python module, with only a subset of the functions within the folder exposed via the module. These exposed functions are known as "public functions" because they are meant to be used by other parts of the application.

Add a new public function called `generate_game` which can be used to easily create a new game with a given number of players.