Extract code from the set_faction_leader function
Add a `set_faction_leader` function to the faction leader function. This function is an extraction from the `select_faction_leader` function, and is designed to allow tests to set the faction leader without it having to provide a related action. Also change the functions called by the `SubmitActionViewSet` to no longer require the game id, because the functions can get this themselves if and when they need it.