Simplify WS messages by merging 2 operation types
Merge the `update` and `create` WebSocket message operation types into one called `create`, leaving just `create` and `destroy` as the only remaining message operation types used in the app. This change was made because I realized that the implementations of the `update` and `create` operations in the frontend were almost identical, so they could be merged. The change reduces the complexity of action-specific code relating to the issuance of and response to WebSocket messages.