Add test coverage for the mortality phase (#346)
* Improve typing in game_starter.py

* Move the mortality phase setup code

Move the mortality phase setup code from the faction leader helper to the mortality phase helper script.

* Extract code for getting and checking actions

Extract the test code for getting and checking actions into a new helper function and make a start on the mortality phase test. The new mortality phase test and the existing faction phase test use this helper function.

* Add action submission to the mortality phase test

Add action submission to the mortality phase test, as well as a check after to ensure that the actions get marked as completed.

* Add a draw mortality chit test

Add a test that covers the function that's used to draw mortality chits.

* Change how random seeds work in tests

Adjusted some of the test expectations due to modifications in the usage of the seed for random number generation. Previously, the seed was set within the function under test, causing the sequence of random numbers to reset with each function call. Now, the seed is set outside the function in the calling code, resulting in a continuous sequence of random numbers across multiple function calls. This change in the seed's scope and timing of initialization leads to a different sequence of random numbers within the function, thus altering the function's output and necessitating updates to the test expectations.

* Add a resolve_mortality function and test coverage

Extract some code from the `face_mortality` function to create a new function called `resolve_mortality`. This function can take chit code overrides, which is useful for testing because tests can now specify exactly who should die without using and built in `random` functions. Add a second test to the `MortalityPhaseTests` test case that checks the `resolve_mortality` function using a chit code override.

Also included in this commit are improvements to test doc comments and Python typing across various functions.

* Extract code from the set_faction_leader function

Add a `set_faction_leader` function to the faction leader function. This function is an extraction from the `select_faction_leader` function, and is designed to allow tests to set the faction leader without it having to provide a related action. Also change the functions called by the `SubmitActionViewSet` to no longer require the game id, because the functions can get this themselves if and when they need it.

* Simplify Django model string representations

Remove the `__str__` methods from all Django models, because although the detailed string representations are useful in the admin site, they're overwhelming when seen in tests. The default string representation (e.g. `Faction object (3)`) is preferred when debugging tests.

* Stop storing old actions in frontend and backend

Stop the frontend from treating potential and completed actions as two separate collections. The frontend now has one collection containing all actions related to the current step. This is simpler than having two collections. In the backend tests, there is now more monitoring of actions, and actions from previous steps are expected to be deleted because there's no point in storing them. Add a backend helper function called `delete_old_actions` to help with this.

* Simplify WS messages by merging 2 operation types

Merge the `update` and `create` WebSocket message operation types into one called `create`, leaving just `create` and `destroy` as the only remaining message operation types used in the app. This change was made because I realized that the implementations of the `update` and `create` operations in the frontend were almost identical, so they could be merged. The change reduces the complexity of action-specific code relating to the issuance of and response to WebSocket messages.

* Create the `check_all_actions` function

Extract the code used to count how many actions are pending and completed from the faction phase test because this will probably be reused in other tests. This code is now in the `check_all_actions` function.

* Add more action checks to the mortality phase