Improve resolve mortality test coverage (#350)
* Improve resolve mortality test coverage.

Improve test coverage of the resolve mortality phase to improve the chances of regressions being detected in the future. The tests now monitor the behavior of the resolve mortality function when a faction leader, a regular senator, or multiple senators are killed.

* Improve mortality double kill test coverage

Add more assertions to the part of the mortality resolution test where two senators are killed because this improves coverage.