Add test coverage for WebSocket messages (#351)
* Move WebSocket message sending responsibility

Shift responsibility for sending WebSocket messages from each action function to the submit action view, because it is common behavior so shouldn't be repeated. This will also make it easier to setup testing for WebSocket messages.

* Enable test running and debugging in VS Code

Enable test running and debugging through VS Code, using pytest. Also remove Django Debug Toolbar because it was causing issues when running the tests through VS Code, and I haven't really used it since I added it.

* Improve the Senator admin page

Improve the Senator admin page to improve developer experience.

* Add WebSocket mortality phase test coverage

Add coverage for WebSocket messages in the mortality phase tests. The test only checks that the correct number of messages are produced, not what the messages contain, which isn't ideal but it's better than nothing.

* Fix Faction WebSocket message processing bug

Allow the frontend to notice Faction updates in WebSocket messages. It should have been able to do this already, but it had been missed. This change fixes the bug where faction rank changes weren't showing up in the frontend unless the page was refreshed.