Add the Prior Consul title (#354)
* Improve some Django admin pages

* Make the backend generate a Prior Consul title

Make the backend generate a Prior Consul title for the Temporary Rome Consul during game setup, and ensure that it works by adding test coverage.

* Swap the FA time icon for a home made one

Swap the Font Awesome time icon for a home made one. Like the old one, the new icon also shows an hourglass.

* Add prior consul to the frontend

Add prior consul to the frontend, including an icon and detail section content for the title.

* Add and use the `TitleList` component

Add the `TitleList` component for generalizing the logic associated with listing a senator's titles. Use this component in the senator detail content and the senator list to show information in the empty space.

* Fix some frontend build errors