Add situations and secrets (#372)
* Add `Situation` and `Secret` models

Add the `Situation` and `Secret` models, which represent cards in the deck and the player hands respectively.

* Fix name field length for secret and situation

* Make game starter create situations and secrets

* Migrate existing games to have new model instances

Migrate existing games to have correctly assigned instances of the new `Situation` and `Secret` models. Fix some mistakes in these models, such as the type choices, and improve their respective admin pages.

* Expose secrets on two API endpoints

Expose secrets on two API endpoints: private for a player to read their own secrets, and public for a player to partially read other players' secrets (name and type are hidden).

* Add basic frontend representation of secrets

* Improve the appearance of the secret list

* Add secrets icon

Add a secrets icon (replacing the votes icon used to temporarily represent secrets in the faction details component) and add secrets as an attribute to the faction list item component. The icon depicts a simple red masquerade mask.

* Fix missing dependency and key prop issues