Improve the faction part of the meta section
Change the "Playing as the X Faction" part of the meta section to look better and add two faction attributes here: votes and secrets.

Create a new component `AttributeFlex`, which is similar to `AttributeGrid` except it's a flex rather than a grid. This introduces no new stuff, but is an extraction of some code that was already in `FactionListItem`. The purpose of the extraction is for use in the `MetaSection` in addition to the existing use in `FactionListItem`.