UI improvements (#377)
* Change code order

Change the code order to match the order of items as they appear in the UI from left to right.

* Add "Your faction has X secrets" text

* Remove "Secret" from secret items

Remove the word "Secret" from individual secrets shown in the secrets list, since it's already obvious that the item is a secret.

* Improve the faction part of the meta section

Change the "Playing as the X Faction" part of the meta section to look better and add two faction attributes here: votes and secrets.

Create a new component `AttributeFlex`, which is similar to `AttributeGrid` except it's a flex rather than a grid. This introduces no new stuff, but is an extraction of some code that was already in `FactionListItem`. The purpose of the extraction is for use in the `MetaSection` in addition to the existing use in `FactionListItem`.

* Add conditional "Your" to faction description

Make the faction details description start out with "Your Faction..." if the selected faction is controlled by the user.