Add test coverage for War initiation and matching
Add test coverage for War initiation and matching, using the initiation of the 2nd Punic War as an example. Change the format of war situation names, so "1st Punic War" is now "Punic 1", which is easier to parse. Also change the index of the Syrian War to 0 so that the game doesn't consider it to be part of a series.