Allow Situations to be initiated (#394)
* Add the New Family situation

* Add test coverage for the New Family sitch

* Add the War model

* Make 1st Punic War an inactive war at game start

Remove 1st Punic War from situation decks and create it as an inactive war. Also add and use more detailed war JSON data.

* Add naval defeated attribute to war

* Expose Wars on the API

* List existing wars in the UI in a "Warfare" tab

* Correct spelling mistake

* Improve the warfare UI

* Make War situations work

Make War Situations initiable and add images for the various wars.

* Add test coverage for War initiation and matching

Add test coverage for War initiation and matching, using the initiation of the 2nd Punic War as an example. Change the format of war situation names, so "1st Punic War" is now "Punic 1", which is easier to parse. Also change the index of the Syrian War to 0 so that the game doesn't consider it to be part of a series.

* Fix turn index bug

Fix a bug that was stopping the turn number from advancing beyond 1.

* Improve the war status UI

* Add War strength multiplication

* Improve matched war test coverage

* Add Leader attribute data

* Add the Leader model

Add the Leader model to the backend. Leader initiation happens, but it doesn't generate an action log, and the frontend can't handle the Leader instance creation messages.

* Add basic Enemy Leader UI representation

* Split wars by status group

Split Wars by status group (Active and Potential) and put Enemy Leader names next to their respective wars.

* Improve UI appearance of Wars

* Fix idle Enemy Leader UI

* Rename file

* Rename file to target name

* Rename files

* Renames files to target names

* Change from stone gray to neural gray

* Further improve Wars UI

* Improve existing war and leader notifications

* Add/improve Enemy Leader notifications

Add a notification for when an Enemy Leader is matched. Also expand usage of the matched_war notification type so it can be used to communicate that a War has been matched (activated) by an Enemy Leader.

* Add some newlines to file endings

* Fix a matching war and leader bugs

* Add test coverage for Wars after Enemy Leaders

Add test coverage for Wars being initiated after Enemy Leaders, so the Leaders join the Wars.

* Add test coverage for new_enemy_leader situation

* Enable secrets to be gained

Enable secrets to be gained during the forum phase. The client page has to be refreshed to see new secrets because there is currently no secure way to send secrets to clients.

* Improve new secret test coverage

* Fix an invalid HTML warning

Fix an invalid HTML warning caused by a `p` containing a `div`.

* Add a player group WebSocket channel

Add a player group WebSocket channel, and allow the frontend to connect to it. Also add a catch-all WS route to the backend for handling invalid WS URLs. The frontend attempts to establish a connection via an invalid URL before it has access to the player ID.

* Send Secret create messages to client

Use the player group WebSocket channel to communicate Secret create messages to the recipient player's client to update state.

* Allow handling of conflicting Secret instances

Allow the frontend to handle conflicting Secret instances on the frontend (public vs private).