Add final forum phase and game over
Make it possible for the game to end. Once all situations have been initiated, the current (forum) phase is renamed to the "Final Forum" phase. When in the final forum phase, players can't take the initiate situation action, they can only select their faction leader.

Once the final forum phase is complete, the game ends. The place in the UI that normally shows the current turn and phase is replaced with "Game over". The action area disappears because nobody is due to take an action.