Reduce saturation for faction background colors (#438)
* Reduce saturation for some faction colors

Reduce saturation for shade 900 faction colors by 33%. This shade is generally used as a background color on dark mode, so reduced saturation makes sense.

* Update colors.css

Reduce the saturation of the shade 900 purple faction color by an additional 11% (representing a net reduction of 50% instead of 33%). This is because, to my eye, the purple hue has an especially saturated appearance and needs additional desaturation to match the other colors.

* Refactor the changes to make them safer

Refactor the changes from previous commits so that the new colors are only used by factions. This should make the changes a bit safer.