Improve the progress section (mostly refactoring) (#440)
* Rename Notification to ActionLog in a few places

* Extract code from ProgressSection to ActionsArea

Extract the parts of the code for the ProgressSection component that relate to actions, and put them into a new component called ActionsArea. Also move the `latestActions` state into the game context.

* Show more notifications by default

Make the notification list start out with 50 notifications rather than 10.

* Make the ActionsArea box have a fixed size

Make the box inside the ActionsArea component have a fixed size, so that it doesn't resize the notification list as the current pending action changes.