Add an action log for when the era ends (#470)
* Add an action log for when the era ends

* Fix the unit tests that deleted all situations

Fix the unit tests that deleted all situations but one, as they were triggering the era ends situation, messing up the test. These tests now set the index for the target situation to 100 to ensure that they are up next.

* Add test coverage for the era ends action log

Add test coverage for the era ends action log and the phase name change.