Create senator action log relations
Make the concession helper create a senator action log relation when a New Concession action log is created. Change the way action logs are retrieved, fixing a bug in the senator History tab. Fix a bug in the game ender that caused the wrong faction to be selected as winner if there was a tie in combined influence.